requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
ImageHash==4.3.1
isodate==0.6.1
lxml==5.1.0
numpy==1.26.3
opencv-python==4.9.0.80
outcome==1.3.0.post0
pillow==10.2.0
pycountry==23.12.11
pycryptodome==3.20.0
PySocks==1.7.1
PyWavelets==1.5.0
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
streamlink==6.5.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
